fix: use export google once
